api/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
# pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
redis==3.5.3
requests==2.25.0
rq==1.6.1
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
